requirements.txt
acl==0.1
alembic==1.7.6
altgraph==0.17
aniso8601==7.0.0
appdirs==1.4.4
arrow==1.0.3
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
automium==0.2.6
automium-web==0.1.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
boto3==1.16.34
botocore==1.19.34
buildozer==1.2.0
bz2file==0.98
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
cheroot==8.5.2
ciseau==1.0.1
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.2
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.24
distlib==0.3.2
Django==3.2.5
django-appconf==1.0.4
django-bootstrap4==3.0.1
django-ckeditor==6.0.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.12.0
django-db-file-storage==0.5.5
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-split-settings==1.0.1
django-storages==1.10.1
Django-Verify-Email==1.0.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.17.1
email-validator==1.1.3
epub==0.5.2
epub-conversion==1.0.15
et-xmlfile==1.1.0
extract-social-media==0.4.0
filelock==3.0.12
fontawesome-free==5.15.3
fpdf==1.7.2
future==0.18.2
gmail-connector==0.2.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==3.1.5
graphql-relay==3.1.0
greenlet==1.1.2
gTTS==2.2.2
gunicorn==20.1.0
html-to-etree==0.2.0
idna==3.2
image==1.5.33
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
jaraco.functools==3.4.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
Kivy-Garden==0.1.4
kiwisolver==1.3.1
legacy==0.1.6
lxml==4.6.3
Mako==1.1.6
MarkupSafe==1.1.1
matplotlib==3.4.2
more-itertools==8.12.0
MouseInfo==0.1.3
numpy==1.20.2
openpyxl==3.0.7
packaging==21.0
panda==0.3.1
pandas==1.3.1
paramiko==2.7.2
pefile==2021.5.24
pexpect==4.8.0
phonenumbers==8.12.28
pil-video==0.0.4
pilkit==2.0
Pillow==8.3.1
pipreqs==0.4.11
platformdirs==2.1.0
playsound==1.2.2
pluggy==0.13.1
poyo==0.5.0
promise==2.3
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
PyAutoGUI==0.9.53
pycparser==2.20
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.9.0
PyGreSQL==5.2.2
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyJWT==2.3.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.27
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.1
python-docx==0.8.11
python-slugify==4.0.1
PyTweening==1.0.3
pytz==2021.1
pywhatkit==5.0
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.26.0
Rx==3.2.0
s3transfer==0.3.3
sh==1.14.2
singledispatch==3.6.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.31
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
turtle==0.0.1
urllib3==1.26.6
virtualenv==20.6.0
w3lib==1.22.0
waitress==2.0.0
web.py==0.62
Werkzeug==1.0.1
whitenoise==5.2.0
wikipedia==1.4.0
WTForms==3.0.1
wxPython==4.1.1
xlrd==2.0.1
XlsxWriter==1.4.4
yarg==0.1.9
zipp==3.7.0
